fix merged wasm unit tests
